fix: move fetching bybit api and forex api to server side to secure headers
